OpenXR module: added support for fixed foveated rendering, setting display refresh rate and the cpu and gpu level.

+++ Modules/VRWrapper/RNVRWindow.h
@@ -61,7 +61,7 @@ namespace RN
 		RNVRAPI virtual void SetFullscreen(bool fullscreen) override {}
 
 		RNVRAPI virtual void SetFixedFoveatedRenderingLevel(uint8 level, bool dynamic) { }
-		RNVRAPI virtual void SetPreferredFramerate(uint32 framerate) { }
+		RNVRAPI virtual void SetPreferredFramerate(float framerate) { }
 		RNVRAPI virtual void SetPerformanceLevel(uint8 cpuLevel, uint8 gpuLevel) { }
 
 		RNVRAPI virtual Vector2 GetSize() const override = 0;
